CulinaryLocation/requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
cryptography==42.0.1
Deprecated==1.2.14
Django==5.0.1
django-bootstrap5==23.4
django-ckeditor==6.7.0
django-firebase==0.0.7
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
firebase-admin==6.4.0
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.1
msgpack==1.0.7
mysqlclient==2.2.1
oauthlib==3.2.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pytz==2023.4
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
